+++ src/components/App.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
-import './App.css'
+import reactLogo from '../assets/react.svg'
+import viteLogo from '../assets/vite.svg'
+import '../styles/App.css'
 
 function App() {
   const [count, setCount] = useState(0)
@@ -22,7 +22,7 @@ function App() {
           count is {count}
         </button>
         <p>
-          Edit <code>src/App.tsx</code> and save to test HMR
+          Edit <code>src/components/App.tsx</code> and save to test HMR
         </p>
       </div>
       <p className="read-the-docs">
